requirement.txt
absl-py==2.1.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
av==13.0.0
attrs==23.2.0
backcall==0.2.0
beartype==0.17.2
beautifulsoup4==4.12.3
black==21.4b2
bleach==6.1.0
blinker==1.8.2
blosc2==2.7.0
botocore==1.34.54
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
cityscapesScripts==2.2.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.2
dacite==1.8.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
einops==0.7.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.14.0
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
glfw==2.7.0
gradio==3.42.0
gradio_client==0.5.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.2
imageio-ffmpeg==0.5.1
imath==0.0.2
imgviz==1.7.5
importlib_metadata==8.0.0
importlib_resources==6.4.0
infinibatch==0.1.1
iopath==0.1.9
ipython==8.12.3
ipywidgets==8.1.3
isort==5.13.2
itsdangerous==2.2.0
jaxtyping==0.2.28
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.4
lightning==2.2.1
lightning-utilities==0.11.0
lit==18.1.8
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmcls==0.25.0
model-index==0.1.11
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mup==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.1
open3d==0.18.0
opencv-python==4.8.1.78
opendatalab==0.0.10
OpenEXR==3.2.4
openxlab==0.1.1
ordered-set==4.1.0
orjson==3.10.6
oss2==2.17.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pascal-voc-writer==0.1.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.22.0
plyfile==1.0.3
portalocker==2.10.0
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pydot==2.0.0
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytz==2023.4
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.10.3
requests==2.28.2
retrying==1.3.4
rich==13.4.2
rpds-py==0.18.1
ruff==0.3.3
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tables==3.9.2
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorbox==0.0.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tiktoken==0.7.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torchmetrics==1.3.2
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
triton==2.0.0
typeguard==2.13.3
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.13
uvicorn==0.30.1
uvloop==0.19.0
vision-datasets==0.2.2
wandb==0.15.12
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xformers==0.0.21
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
yarl==1.9.4
zipp==3.19.2
huggingface-hub==0.24.0
tokenizers
timm
transformers
viser
flow_vis_torch @ git+https://github.moeyy.xyz/https://github.com/ChristophReich1996/Optical-Flow-Visualization-PyTorch@9177370c7c00b4b7dbe4deda6fed734fdff48b2c